refactor(repo): fix test config, remove pointless helpers, fix IDE intellisense
- Fix broken `@test` alias in core vitest.config.ts and tsconfig.json
  (`./test` → `./src/test` after prior refactor deleted the old dir)
- Update core test imports from `@test/core-utils.js` to `@test/index.js`
- Delete `tests/helpers/core-utils.ts` (was just re-exporting `@kidd-cli/core/test`)
- Delete `packages/core/test/core-utils.ts` (stale file from prior refactor)
- Create `tests/helpers/index.ts` barrel for actual unique helpers (mock-oauth-server)
- Integration tests import core test utils directly from `@kidd-cli/core/test`
- Remove `composite: true` from `tests/tsconfig.json` (had `noEmit: true`, was
  never typechecked by CI, composite only added IDE restrictions)
- Remove dead `{ "path": "tests" }` reference from root `tsconfig.json`
- Add `@kidd-cli/core` and `@kidd-cli/core/test` path mappings to tests tsconfig
- Add `@clack/prompts` as root devDep for integration test imports
