+++ package.json
@@ -26,6 +26,7 @@
   },
   "devDependencies": {
     "@changesets/cli": "^2.30.0",
+    "@clack/prompts": "^1.1.0",
     "@types/node": "catalog:",
     "@typescript/native-preview": "7.0.0-dev.20260315.1",
     "@vitest/coverage-v8": "catalog:",
