+++ packages/cli/CHANGELOG.md
@@ -20,7 +20,6 @@
 ### Patch Changes
 
 - 0d0c61f: Comprehensive code review cleanup and refactoring across all packages:
-
   - Reorganize bundler into `build/`, `compile/`, `autoloader/`, `config/` subdirectories
   - Split core `types.ts` into `types/utility`, `types/middleware`, `types/command`, `types/cli`
   - Create shared `tsdown.base.mjs` config for all packages
@@ -84,13 +83,11 @@
 - 6d8889a: Add `ConfigType` utility type and `CliConfig` augmentation interface for typed `ctx.config`.
 
   **@kidd-cli/core:**
-
   - Add `ConfigType<TSchema>` utility type to derive `CliConfig` from a Zod schema
   - Rename `KiddConfig` augmentation interface to `CliConfig` to avoid confusion with the build config type in `@kidd-cli/config`
   - Export `CliConfig` and `ConfigType` from `@kidd-cli/core`
 
   **@kidd-cli/cli:**
-
   - Add `--config` flag to `kidd init` to scaffold config schema setup during project creation
   - Add `kidd add config` command to scaffold config into existing projects
   - Scaffolded config includes Zod schema with `ConfigType` module augmentation wiring
