+++ packages/core/CHANGELOG.md
@@ -11,7 +11,6 @@
 ### Minor Changes
 
 - e4ebe22: Add hidden, deprecated, and group support for commands and flags
-
   - Add `Resolvable<T>` utility type (`T | (() => T)`) for values resolved at registration time
   - Add `hidden` field on `CommandDef` and `YargsArgDef` to omit commands/flags from help output
   - Add `deprecated` field on `CommandDef` and `YargsArgDef` to show deprecation notices
@@ -24,7 +23,6 @@
 ### Patch Changes
 
 - 0d0c61f: Comprehensive code review cleanup and refactoring across all packages:
-
   - Reorganize bundler into `build/`, `compile/`, `autoloader/`, `config/` subdirectories
   - Split core `types.ts` into `types/utility`, `types/middleware`, `types/command`, `types/cli`
   - Create shared `tsdown.base.mjs` config for all packages
@@ -56,7 +54,6 @@
 ### Minor Changes
 
 - b1c8e9e: Refactor config client to use c12 for all config file resolution
-
   - Support `name.config.*` patterns (TS, JS, JSON, JSONC, YAML, TOML) via c12
   - Support `name.*` short-form patterns for data formats only (JSON, JSONC, YAML, TOML)
   - Long form (`name.config.*`) takes priority over short form (`name.*`)
@@ -87,13 +84,11 @@
 - 6d8889a: Add `ConfigType` utility type and `CliConfig` augmentation interface for typed `ctx.config`.
 
   **@kidd-cli/core:**
-
   - Add `ConfigType<TSchema>` utility type to derive `CliConfig` from a Zod schema
   - Rename `KiddConfig` augmentation interface to `CliConfig` to avoid confusion with the build config type in `@kidd-cli/config`
   - Export `CliConfig` and `ConfigType` from `@kidd-cli/core`
 
   **@kidd-cli/cli:**
-
   - Add `--config` flag to `kidd init` to scaffold config schema setup during project creation
   - Add `kidd add config` command to scaffold config into existing projects
   - Scaffolded config includes Zod schema with `ConfigType` module augmentation wiring
@@ -101,14 +96,12 @@
 - 70deba8: Redesign output API: replace `ctx.output` with `ctx.format` and add styled logger methods.
 
   **Breaking changes:**
-
   - Remove `ctx.output` from the Context (replaced by `ctx.format` and `ctx.logger`)
   - Rename `SYMBOLS`/`Symbols` to `GLYPHS`/`Glyphs`
   - Rename format types: `ResultInput` to `CheckInput`, `DiagnosticInput` to `FindingInput`, `SummaryInput` to `TallyInput`, `TallySummaryInput` to `TallyBlockInput`, `InlineSummaryInput` to `TallyInlineInput`, `ResultStatus` to `CheckStatus`, `DiagnosticSeverity` to `FindingSeverity`
   - Rename format functions: `formatResult` to `formatCheck`, `formatDiagnostic` to `formatFinding`, `formatSummary` to `formatTally`
 
   **New features:**
-
   - Add `ctx.format.json(data)` and `ctx.format.table(rows)` — pure string formatters (no I/O)
   - Add `ctx.logger.check(input)` — write a pass/fail/warn/skip/fix row (vitest style)
   - Add `ctx.logger.finding(input)` — write a full finding with optional code frame (oxlint style)
@@ -118,13 +111,13 @@
 
   ```ts
   // Before
-  ctx.output.result(input); // → ctx.logger.check(input)
-  ctx.output.diagnostic(input); // → ctx.logger.finding(input)
-  ctx.output.summary(input); // → ctx.logger.tally(input)
-  ctx.output.write(data); // → process.stdout.write(ctx.format.json(data))
-  ctx.output.table(rows); // → process.stdout.write(ctx.format.table(rows))
-  ctx.output.raw(text); // → ctx.logger.print(text)
-  ctx.output.markdown(text); // → ctx.logger.print(text)
+  ctx.output.result(input) // → ctx.logger.check(input)
+  ctx.output.diagnostic(input) // → ctx.logger.finding(input)
+  ctx.output.summary(input) // → ctx.logger.tally(input)
+  ctx.output.write(data) // → process.stdout.write(ctx.format.json(data))
+  ctx.output.table(rows) // → process.stdout.write(ctx.format.table(rows))
+  ctx.output.raw(text) // → ctx.logger.print(text)
+  ctx.output.markdown(text) // → ctx.logger.print(text)
   ```
 
 ### Patch Changes
@@ -188,7 +181,6 @@
   **Auth HTTP integration:** `auth({ http: { baseUrl, namespace } })` creates authenticated HTTP clients with automatic credential header injection. Supports single or multiple clients via an array.
 
   **Breaking changes:**
-
   - `http()` no longer auto-reads `ctx.auth.credential()`. Use `auth({ http })` for authenticated clients or pass `headers` explicitly.
   - `HttpOptions.defaultHeaders` renamed to `headers` and now accepts a function `(ctx) => Record<string, string>` in addition to a static record.
 
@@ -196,9 +188,9 @@
 
   ```ts
   middleware: [
-    auth({ resolvers: [{ source: "env" }] }),
-    http({ baseUrl: "https://api.example.com", namespace: "api" }),
-  ];
+    auth({ resolvers: [{ source: 'env' }] }),
+    http({ baseUrl: 'https://api.example.com', namespace: 'api' }),
+  ]
   ```
 
   After:
@@ -207,9 +199,9 @@
   middleware: [
     auth({
       resolvers: [auth.env()],
-      http: { baseUrl: "https://api.example.com", namespace: "api" },
+      http: { baseUrl: 'https://api.example.com', namespace: 'api' },
     }),
-  ];
+  ]
   ```
 
 - f48ad38: Replace non-standard OAuth flow with spec-compliant PKCE (RFC 7636) and add Device Authorization Grant (RFC 8628)
