+++ packages/core/src/runtime/args.test.ts
@@ -1,4 +1,4 @@
-import { setArgv, runTestCli, setupTestLifecycle } from '@test/core-utils.js'
+import { setArgv, runTestCli, setupTestLifecycle } from '@test/index.js'
 import { describe, expect, it, vi } from 'vitest'
 import { z } from 'zod'
