+++ packages/core/tsconfig.json
@@ -10,7 +10,7 @@
     "rootDir": "./src",
     "paths": {
       "@/*": ["./src/*"],
-      "@test/*": ["./test/*"]
+      "@test/*": ["./src/test/*"]
     }
   },
   "include": ["src"],
