+++ tests/integration/auth-oauth-pkce.test.ts
@@ -18,8 +18,8 @@ import { execFile } from 'node:child_process'
 
 import { resolveFromOAuth } from '@/middleware/auth/strategies/oauth.js'
 
-import { createMockOAuthServer } from '../helpers/mock-oauth-server.js'
-import type { MockOAuthServer } from '../helpers/mock-oauth-server.js'
+import { createMockOAuthServer } from '../helpers/index.js'
+import type { MockOAuthServer } from '../helpers/index.js'
 
 const KNOWN_STATE = Buffer.from('a'.repeat(32)).toString('hex')
 const KNOWN_VERIFIER = Buffer.from('a'.repeat(32)).toString('base64url')
