+++ tests/integration/simple-list.test.ts
@@ -1,3 +1,4 @@
+import { runTestCli, setArgv, setupTestLifecycle } from '@kidd-cli/core/test'
 import { describe, expect, it, vi } from 'vitest'
 
 import { command } from '@/command.js'
@@ -6,7 +7,6 @@ import type { Context } from '@/context/types.js'
 import type { CommandMap } from '@/types.js'
 
 import listCommand from '../../examples/simple/commands/list.js'
-import { runTestCli, setArgv, setupTestLifecycle } from '../helpers/core-utils.js'
 
 const mockSpinnerInstance = vi.hoisted(() => ({
   message: vi.fn(),
