+++ tests/tsconfig.json
@@ -1,10 +1,11 @@
 {
   "extends": "../tsconfig.base.json",
   "compilerOptions": {
-    "composite": true,
     "noEmit": true,
     "paths": {
-      "@/*": ["../packages/core/src/*"]
+      "@/*": ["../packages/core/src/*"],
+      "@kidd-cli/core": ["../packages/core/src/index.ts"],
+      "@kidd-cli/core/test": ["../packages/core/src/test/index.ts"]
     }
   },
   "include": ["./**/*.ts"]
