+++ tsconfig.json
@@ -5,7 +5,6 @@
     { "path": "packages/cli" },
     { "path": "packages/config" },
     { "path": "packages/core" },
-    { "path": "packages/utils" },
-    { "path": "tests" }
+    { "path": "packages/utils" }
   ]
 }
